Improved pattern matching for modules and functions
Module-level parsing now ignores any `if`, `for` or `while` statements
that for whatever reason may appear in the module itself (the likely use
case being the inclusion of an `if __name__ == '__main__'` check).

Function-level parsing now correctly handles markup for return values
containing square brackets as part of the syntactic sugar for container
`typing` includes like List and Tuple.